requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.7
djangorestframework==3.15.2
filelock==3.17.0
platformdirs==4.3.6
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.29.1
